requirements.txt
discord.py==2.3.2
Pillow==10.1.0
PyMySQL[ed25519]==1.1.0
yookassa==3.0.0
aiohttp==3.9.1
aiodns==3.1.1
fastapi==0.105.0
uvicorn[standard]==0.25.0
MinePI==0.5.0
aio-mc-rcon==3.2.1
python-dateutil==2.8.2
aiofiles==23.2.1
dynaconf==3.2.4